+++ ant-boot-starters/ant-boot-starter-env/pom.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0"?>
+<project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
+         xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>org.antframework.boot</groupId>
+        <artifactId>ant-boot-starters</artifactId>
+        <version>1.3.0.RELEASE</version>
+    </parent>
+
+    <artifactId>ant-boot-starter-env</artifactId>
+
+    <dependencies>
+    </dependencies>
+</project>
