+++ ant-boot-starters/ant-boot-starter-env/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,2 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+  org.antframework.boot.env.boot.EnvAutoConfiguration
