+++ ant-boot-starters/ant-boot-starter-lang/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,2 @@
+org.springframework.boot.SpringApplicationRunListener=\
+  org.antframework.boot.lang.boot.LangApplicationRunListener
